+++ lib/rest/proxy/v1/service/session.d.ts
@@ -33,7 +33,7 @@ declare function SessionList(version: V1, serviceSid: string): SessionListInstan
  * Options to pass to update
  *
  * @property dateExpiry - The ISO 8601 date when the Session should expire
- * @property failOnParticipantConflict - Opt-in to enable Proxy to return 400 on detected conflict on re-open request.
+ * @property failOnParticipantConflict - An experimental flag that instructs Proxy to return 400 instead of 200 when it detects that conflicts would result from re-open requests.
  * @property status - The new status of the resource
  * @property ttl - When the session will expire
  */
@@ -174,6 +174,7 @@ interface SessionListInstance {
  * Options to pass to create
  *
  * @property dateExpiry - The ISO 8601 date when the Session should expire
+ * @property failOnParticipantConflict - An experimental flag that instructs Proxy to reject a Session create request when it detects a Participant conflict.
  * @property mode - The Mode of the Session
  * @property participants - The Participant objects to include in the new session
  * @property status - Session status
@@ -182,6 +183,7 @@ interface SessionListInstance {
  */
 interface SessionListInstanceCreateOptions {
   dateExpiry?: Date;
+  failOnParticipantConflict?: boolean;
   mode?: SessionMode;
   participants?: object | object[];
   status?: SessionStatus;